kvm(x86): introduce kvm_access_guest_gfn_{start,end}
This pair of functions forms a "context manager" to be used around
accesses to (potentially private) guest memory from within KVM (such as
for instruction fetch during MMIO emulation). They temporarily reinsert
the pfn backing the given gfn into the kernel's direct map, so that we
can access it using a khva.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>