fix: reinsert kvm-clock page back into direct map
A KVM guest using kvm-clock as its clocksource communicates with KVM via
a designated page in guest memory via with information for time
sychronization is exchanged. KVM accesses this page via the direct map
(for example in `kvm_setup_guest_pvclock`) in a context where irqs are
disabled, so we cannot use `get_user`/`put_user`. Thus, simply reinsert
the page that the guest communicates to KVM via MSRs to be the kvm-clock
page back into the direct map.

A real solution here would be to have the guest issue
`KVM_HC_MAP_GPA_RANGE` on the kvm-clock page, setting it to shared, and
then KVM can deal with it appropriately in the hypercall handler.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>