kvm: x86: support walking guest page tables in gmem
Update the logic in paging_tmpl.h to work with guest_private memory. If
KVM cannot access gmem and the guest's page tables are in gfns marked as
private, then error out.

We let the paging code access gmem by making it use gfn_to_pfn_caches,
which are already gmem aware, and will later also handle on-demand
mapping of gmem once it supports being removed from the direct map. The
only case that now needs some special handling is page tables in
read-only memslots, as gfn_to_pfn_caches cannot be used for readonly
memory. In this case, we simply use kvm_vcpu_read_guest (which is also
gmem aware), as there is no need to cache the gfn->pfn translation in
this case (as there will not be a second access to the pte. We do not
set the accessed bit for read-only ptes).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>