kvm: gmem: Temporarily restore direct map entries when needed
If KVM_GMEM_NO_DIRECT_MAP is set, and KVM tries to internally access
guest-private memory inside kvm_{read,write}_guest, or via a
gfn_to_pfn_cache, temporarily restore the direct map entry.

To avoid race conditions between two threads restoring or zapping direct
map entries for the same page and potentially interfering with each
other (e.g. unfortune interweavings of map->read->unmap in the form of
map(A)->map(B)->read(A)->unmap(A)->read(B) [BOOM]), we uphold the
following invariant:

- Only a single gfn_to_pfn_cache can exist for any given pfn, and
- All non-gfn_to_pfn_cache code paths that temporarily restore direct
  map entries complete the entire map->access->unmap critical section
  while holding the folio lock.

To remember whether a given folio currently has a direct map entry, we
use the PG_private flag. If this flag is set, then the folio is removed
from the direct map, otherwise it is present in the direct map.
Modifications of this flag, together with the corresponding direct map
manipulations, must happen while holding the folio's lock.

A gfn_to_pfn_cache cannot hold the folio lock for its entire lifetime,
so it operates as follows: In gpc_map, under folio lock, restore the
direct map entry and set PG_private to 0. In gpc_unmap, zap the direct
map entry again and set PG_private back to 1.

If inside gpc_map we find a folio that has PG_private set to 0, then we
know that another gfn_to_pfn_cache is currently active for the given pfn
(as this is the only scenario in which PG_private can be 0 without the
folio lock being held), and so we return -EINVAL.

The only other interesting scenario is then if kvm_{read,write}_guest is
called for a gfn whose translation is currently cached inside a
gfn_to_pfn_cache. In this case, kvm_{read,write}_guest notices that
PG_private is 0 and skips all direct map manipulations. Since it is
holding the folio lock, we can be sure that gpc_unmap cannot zap the
direct map entries while kvm_{read,write}_guest still needs them.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>