kvm: gmem: Refcount internal accesses to gmem
Currently, if KVM_GMEM_NO_DIRECT_MAP is enabled and KVM wants to
internall access a gmem folio, KVM needs to reinsert the folio into the
direct map, and hold the folio lock until KVM is done using the folio
(and the folio is removed from the direct map again).

This means that long-term reinsertion into the direct map, and
concurrent accesses to the same gmem folio are currently impossible.
These are needed however for data structures of paravirtual devices,
such as kvm-clock, which are shared via the guest via guest memory
pages (and multiple vCPUs can put their kvm-clock data into the same
guest page).

Thus, introduce the concept of a "sharing refcount", which gets
incremented on every call to kvm_gmem_get_pfn with
KVM_GMEM_GET_PFN_SHARED set. Direct map manipulations are only done when
the first refcount is grabbed (direct map entries are restored), or when
the last reference goes away (direct map entries are removed). While
holding a sharing reference, the folio lock may be dropped, as the
refcounting ensures that the direct map entry will not be removed as
long as at least one reference is held. However, whoever is holding the
reference will need to listen to gmem invalidation events, and drop its
reference upon receiving one (for example, because the page is in the
process of being fallocated away).

Leaking a reference is again not a fatal issue, although we still detect
and WARN_ONCE if a folio is released while sharing references are held
(as this indicates a KVM bug).

Since refcount_t does not play nicely with references dropping to 0 and
later being raised again, we use a refcount of 1 to mean "no sharing
references held anywhere, folio not in direct map".

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>