mm: reject guest_memfd folios whereever secretmem is rejected
guest_memfd and secretmem (memfd_secret) are similar in the sense that
they represent an area of memory that should not be accessed except from
a specific context (userspace in the case of memfd_secret, KVM guests in
the case of guest_memfd). Furthermore, as part of this patch series,
userspace will be able to request guest_memfd folios to be direct map
removed (exactly like secretmem). In light of this, they will need to be
rejected from all locations that are known to not be able to deal with
direct map removed folios.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>