KVM: guest_memfd: Add flag to remove from direct map
Add KVM_GMEM_NO_DIRECT_MAP flag for KVM_CREATE_GUEST_MEMFD ioctl. When
enabled, guest_memfd folios will be removed from the direct map when
marked as "prepared", with direct map entries only restored when the
folios are freed.

Note that this flag causes removal of direct map entries for all
guest_memfd folios independent of whether they are "shared" or "private"
(although current guest_memfd only supports either all folios in the
"shared" state, or all folios in the "private" state if
!IS_ENABLED(CONFIG_KVM_GMEM_SHARED_MEM)). The usecase for this are a
special type of non-CoCo VM where host userspace is trusted to have
access to all of guest memory, but we still wish to mitigate
Spectre-style transient execution attacks through the host kernel's
direct map.

Note that KVM retains access to guest memory via userspace
mappings of guest_memfd, which are reflected back into KVM's memslots
via userspace_addr. This is needed for things like MMIO emulation on
x86_64 to work. Previous iterations attempted to instead have KVM
temporarily restore direct map entries whenever such an access to guest
memory was needed, but this turned out to have a significant performance
impact, as well as additional complexity due to needing to refcount
direct map reinsertion operations, and making them play nicely with gmem
truncations.

This iteration also doesn't have KVM perform TLB flushes after direct
map manipulations. This is because TLB flushes resulted in a up to 40x
elongation of page faults in guest_memfd (scaling with the number of CPU
cores). On the one hand, TLB flushes are not needed for functional
correctness (the virt->phys mapping technically stays "correct",  the
kernel should simply to not it for a while), so this is a correct
optimization to make. On the other hand, it means that the desired
protection from Spectre-style attacks is not perfect, as an attacker
could try to prevent a stale TLB entry from getting evicted .

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>