KVM: selftests: Add guest_memfd based vm_mem_backing_src_types
Allow selftests to configure their memslots such that userspace_addr is
set to a MAP_SHARED mapping of the guest_memfd that's associated with
the memslot. This setup is the configuration for non-CoCo VMs, where all
guest memory is backed by a guest_memfd whose folios are all marked
shared, but KVM is still able to access guest memory to provide
functionality such as MMIO emulation on x86.

Add backing types for normal guest_memfd, as well as direct map removed
guest_memfd.

If KVM_CAP_MEMORY_ATTRIBUTES is available, explicitly set gmem-enabled
memslots to private, as otherwise guest page faults will be resolved by
GUP-ing the guest_memfd VMA (instead of using the special VMA-less
guest_memfd fault code in the KVM MMU), but this is not always supported
(e.g. if direct map entries are not available).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>